--- miscellaneous/metaDataRemover/config.toml
@@ -1,5 +1,0 @@
-ProxyUrl = ""
-
-TokensToDeletePerTransaction = 0
-
-GasLimit = 0
