+++ tokensRemover/metaDataRemover/config/config.go
@@ -10,6 +10,7 @@ type ContextFlagsMetaDataRemover struct {
 	Pems    string
 }
 
+// Config holds the config for meta data remover tool
 type Config struct {
 	ProxyUrl                     string `toml:"ProxyUrl"`
 	TokensToDeletePerTransaction uint64 `toml:"TokensToDeletePerTransaction"`
