+++ tokensRemover/metaDataRemover/txCreator.go
@@ -15,32 +15,8 @@ import (
 	"github.com/ElrondNetwork/elrond-sdk-erdgo/core"
 	"github.com/ElrondNetwork/elrond-sdk-erdgo/data"
 	"github.com/ElrondNetwork/elrond-sdk-erdgo/interactors"
-	"github.com/ElrondNetwork/elrond-tools-go/miscellaneous/metaDataRemover/config"
+	"github.com/ElrondNetwork/elrond-tools-go/tokensRemover/metaDataRemover/config"
 )
-
-type pkAddress struct {
-	privateKey []byte
-	address    core.AddressHandler
-}
-
-func getPrivateKeyAndAddress(pemFile string) (*pkAddress, error) {
-	w := interactors.NewWallet()
-	privateKey, err := w.LoadPrivateKeyFromPemFile(pemFile)
-	if err != nil {
-		return nil, err
-	}
-
-	address, err := w.GetAddressFromPrivateKey(privateKey)
-	if err != nil {
-		return nil, err
-
-	}
-
-	return &pkAddress{
-		privateKey: privateKey,
-		address:    address,
-	}, nil
-}
 
 func createShardTxs(
 	outFile string,
