+++ tokensRemover/metaDataRemover/txDataFormater_test.go
